+++ app/src/main/java/ubr/persanal/movieapp/domain/usecase/GetActorsUseCase.kt
@@ -5,6 +5,6 @@ import ubr.persanal.movieapp.domain.model.MovieActorsDto
 import ubr.persanal.movieapp.util.ResourceUI
 
 interface GetActorsUseCase {
-    suspend fun invoke(movieId: Int) : Flow<ResourceUI<MovieActorsDto>>
+    suspend fun invoke(movieId: Long) : Flow<ResourceUI<MovieActorsDto>>
 
 }
